# fix: Implement performance optimizations for audio processing