--- a/.gitignore
+++ b/.gitignore
@@ -188,8 +188,7 @@ ENV/
 env.bak/
 venv.bak/
 
-# User-specific scripts
-main.sh
+
 
 # Debug and analysis scripts (temporary)
 scripts/debug/