--- a/dashboard/__init__.py
+++ b/dashboard/__init__.py
@@ -1,0 +1,15 @@
+"""Audio enhancement monitoring dashboard package."""
+
+from .enhancement_dashboard import EnhancementDashboard
+from .metrics_collector import MetricsCollector
+from .comparison_ui import ComparisonAnalyzer
+from .batch_monitor import BatchProcessingMonitor
+from .config_ui import ConfigurationUI
+
+__all__ = [
+    'EnhancementDashboard',
+    'MetricsCollector',
+    'ComparisonAnalyzer',
+    'BatchProcessingMonitor',
+    'ConfigurationUI'
+]