--- a/processors/audio_enhancement/__init__.py
+++ b/processors/audio_enhancement/__init__.py
@@ -1,0 +1,8 @@
+"""
+Audio Enhancement Module
+Provides noise reduction and voice clarity enhancement for audio processing.
+"""
+
+from .core import AudioEnhancer
+
+__all__ = ['AudioEnhancer']