… CHANGELOG.md
… CLAUDE.md
… PARALLEL_DEVELOPMENT.md
… PARALLEL_EXECUTION_SUMMARY.md
… README.md
… RESULTS.md
… WORKTREE_README.md
… dashboard/comparison_ui.py
… dashboard/config_ui.py
… dashboard/configuration_ui.py
… dashboard/enhancement_dashboard.py
… dashboard/metrics_collector.py
… dashboard/static/index.html
… dashboard/static/style.css
… docs/append_id_fix_summary.md
… docs/change-log.md
… docs/checkpoint_duplication_fix_summary.md
… docs/code_cleanup_2025.md
… docs/code_quality_report_2025.md
… docs/hf_viewer_fix_summary.md
… docs/omegaconf_compatibility_fix_summary.md
… docs/secondary_speaker_removal_fix_summary.md
… docs/secondary_speaker_removal_improvements.md
… docs/speaker_clustering_fix_summary.md
… docs/speaker_id_resume_fix_summary.md
… docs/tdd_revert_summary.md
… examples/test_simple_streaming.py
… main.sh
… main_append_test.sh
… monitoring/__init__.py